backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
mistralai==0.1.8
python-multipart==0.0.20
python-dotenv==1.0.1

